% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_precisely.R
\name{launch_precisely_app}
\alias{launch_precisely_app}
\title{Launch precisely Shiny app}
\usage{
launch_precisely_app()
}
\description{
\code{launch_precisely_app()} launches a Shiny app to calculate and plot
precision, sample size, and upper limit calculations.
}
