% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_precisely.R
\name{theme_precisely}
\alias{theme_precisely}
\title{Minimalist themes for precision plots}
\usage{
theme_precisely(base_size = 14, base_family = "", ...)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{...}{additional arguments passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}}
}
\description{
Minimalist themes for precision plots
}
